# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArcMeme.fun API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TokenMarketType =
  (typeof TokenMarketType)[keyof typeof TokenMarketType];

export const TokenMarketType = {
  unlisted: "unlisted",
  amm_pool: "amm_pool",
} as const;

export interface Token {
  id: string;
  name: string;
  ticker: string;
  price: number;
  marketCap: number;
  volume24h: number;
  change24h: number;
  description: string;
  createdAt: string;
  creatorAddress: string;
  logoColor: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  contractAddress?: string | null;
  marketType?: TokenMarketType;
  /** @nullable */
  pairAddress?: string | null;
  /** @nullable */
  routerAddress?: string | null;
  totalSupply: number;
  holders: number;
  txCount: number;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  twitter?: string | null;
  /** @nullable */
  telegram?: string | null;
}

export interface TokenInput {
  /**
   * @minLength 1
   * @maxLength 50
   */
  name: string;
  /**
   * @minLength 1
   * @maxLength 10
   */
  ticker: string;
  description: string;
  website?: string;
  twitter?: string;
  telegram?: string;
  logoColor?: string;
  logoImage?: string;
  totalSupply?: number;
  contractAddress?: string;
  creatorAddress?: string;
}

export type TokenMarketInputMarketType =
  (typeof TokenMarketInputMarketType)[keyof typeof TokenMarketInputMarketType];

export const TokenMarketInputMarketType = {
  unlisted: "unlisted",
  amm_pool: "amm_pool",
} as const;

export interface TokenMarketInput {
  marketType: TokenMarketInputMarketType;
  /** @nullable */
  pairAddress: string | null;
  /** @nullable */
  routerAddress: string | null;
}

export type TradeSide = (typeof TradeSide)[keyof typeof TradeSide];

export const TradeSide = {
  buy: "buy",
  sell: "sell",
} as const;

export interface Trade {
  id: string;
  tokenId: string;
  pairAddress: string;
  txHash: string;
  logIndex: number;
  blockNumber: number;
  side: TradeSide;
  tokenAmount: number;
  wusdcAmount: number;
  executionPrice: number;
  traderAddress: string;
  timestamp: string;
}

export interface ChartPoint {
  timestamp: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export interface Candle {
  time: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export interface PlatformStats {
  totalTokens: number;
  totalVolume24h: number;
  totalMarketCap: number;
  activeTraders: number;
  tokensLaunched24h: number;
}

export type ListTokensParams = {
  sort?: ListTokensSort;
  limit?: number;
};

export type ListTokensSort =
  (typeof ListTokensSort)[keyof typeof ListTokensSort];

export const ListTokensSort = {
  trending: "trending",
  newest: "newest",
  marketCap: "marketCap",
  volume: "volume",
} as const;

export type GetTokenCandlesParams = {
  interval?: GetTokenCandlesInterval;
};

export type GetTokenCandlesInterval =
  (typeof GetTokenCandlesInterval)[keyof typeof GetTokenCandlesInterval];

export const GetTokenCandlesInterval = {
  "1m": "1m",
  "5m": "5m",
  "15m": "15m",
  "1h": "1h",
  "4h": "4h",
} as const;