# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ArcMeme.fun API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tokens
    description: Meme token operations
  - name: stats
    description: Platform statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tokens:
    get:
      operationId: listTokens
      tags: [tokens]
      summary: List all tokens
      parameters:
        - name: sort
          in: query
          schema:
            type: string
            enum: [trending, newest, marketCap, volume]
          required: false
        - name: limit
          in: query
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: List of tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Token"
    post:
      operationId: launchToken
      tags: [tokens]
      summary: Launch a new meme token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenInput"
      responses:
        "201":
          description: Token launched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"

  /tokens/trending:
    get:
      operationId: getTrendingTokens
      tags: [tokens]
      summary: Get trending tokens
      responses:
        "200":
          description: Trending tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Token"

  /tokens/{id}:
    get:
      operationId: getToken
      tags: [tokens]
      summary: Get token by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "404":
          description: Not found

  /tokens/{id}/market:
    patch:
      operationId: updateTokenMarket
      tags: [tokens]
      summary: Update token market metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenMarketInput"
      responses:
        "200":
          description: Token market metadata updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "404":
          description: Not found

  /tokens/{id}/trades:
    get:
      operationId: getTokenTrades
      tags: [tokens]
      summary: Get recent token trades
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Recent trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
        "404":
          description: Not found

  /tokens/{id}/candles:
    get:
      operationId: getTokenCandles
      tags: [tokens]
      summary: Get OHLC candles for a token
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: interval
          in: query
          required: false
          schema:
            type: string
            enum: [1m, 5m, 15m, 1h, 4h]
            default: 1m
      responses:
        "200":
          description: OHLC candles generated from indexed swaps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candle"
        "404":
          description: Not found

  /tokens/{id}/chart:
    get:
      operationId: getTokenChart
      tags: [tokens]
      summary: Get chart data for a token
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chart data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartPoint"

  /stats:
    get:
      operationId: getPlatformStats
      tags: [stats]
      summary: Get overall platform statistics
      responses:
        "200":
          description: Platform statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Token:
      type: object
      required:
        - id
        - name
        - ticker
        - price
        - marketCap
        - volume24h
        - change24h
        - description
        - createdAt
        - creatorAddress
        - logoColor
        - totalSupply
        - holders
        - txCount
      properties:
        id:
          type: string
        name:
          type: string
        ticker:
          type: string
        price:
          type: number
        marketCap:
          type: number
        volume24h:
          type: number
        change24h:
          type: number
        description:
          type: string
        createdAt:
          type: string
        creatorAddress:
          type: string
        logoColor:
          type: string
        logoUrl:
          type: ["string", "null"]
        contractAddress:
          type: ["string", "null"]
        marketType:
          type: string
          enum: [unlisted, amm_pool]
        pairAddress:
          type: ["string", "null"]
        routerAddress:
          type: ["string", "null"]
        totalSupply:
          type: number
        holders:
          type: integer
        txCount:
          type: integer
        website:
          type: ["string", "null"]
        twitter:
          type: ["string", "null"]
        telegram:
          type: ["string", "null"]

    TokenInput:
      type: object
      required:
        - name
        - ticker
        - description
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        ticker:
          type: string
          minLength: 1
          maxLength: 10
        description:
          type: string
        website:
          type: string
        twitter:
          type: string
        telegram:
          type: string
        logoColor:
          type: string
        logoImage:
          type: string
        totalSupply:
          type: number
        contractAddress:
          type: string
        creatorAddress:
          type: string

    TokenMarketInput:
      type: object
      required:
        - marketType
        - pairAddress
        - routerAddress
      properties:
        marketType:
          type: string
          enum: [unlisted, amm_pool]
        pairAddress:
          type: ["string", "null"]
        routerAddress:
          type: ["string", "null"]

    Trade:
      type: object
      required:
        - id
        - tokenId
        - pairAddress
        - txHash
        - logIndex
        - blockNumber
        - side
        - tokenAmount
        - wusdcAmount
        - executionPrice
        - traderAddress
        - timestamp
      properties:
        id:
          type: string
        tokenId:
          type: string
        pairAddress:
          type: string
        txHash:
          type: string
        logIndex:
          type: integer
        blockNumber:
          type: integer
        side:
          type: string
          enum: [buy, sell]
        tokenAmount:
          type: number
        wusdcAmount:
          type: number
        executionPrice:
          type: number
        traderAddress:
          type: string
        timestamp:
          type: string

    ChartPoint:
      type: object
      required:
        - timestamp
        - open
        - high
        - low
        - close
        - volume
      properties:
        timestamp:
          type: integer
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number

    Candle:
      type: object
      required:
        - time
        - open
        - high
        - low
        - close
        - volume
      properties:
        time:
          type: integer
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number

    PlatformStats:
      type: object
      required:
        - totalTokens
        - totalVolume24h
        - totalMarketCap
        - activeTraders
        - tokensLaunched24h
      properties:
        totalTokens:
          type: integer
        totalVolume24h:
          type: number
        totalMarketCap:
          type: number
        activeTraders:
          type: integer
        tokensLaunched24h:
          type: integer