# lib/api-zod/src/generated/types/candle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArcMeme.fun API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Candle {
  time: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}