# lib/api-zod/src/generated/types/token.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArcMeme.fun API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TokenMarketType } from "./tokenMarketType";

export interface Token {
  id: string;
  name: string;
  ticker: string;
  price: number;
  marketCap: number;
  volume24h: number;
  change24h: number;
  description: string;
  createdAt: string;
  creatorAddress: string;
  logoColor: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  contractAddress?: string | null;
  marketType?: TokenMarketType;
  /** @nullable */
  pairAddress?: string | null;
  /** @nullable */
  routerAddress?: string | null;
  totalSupply: number;
  holders: number;
  txCount: number;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  twitter?: string | null;
  /** @nullable */
  telegram?: string | null;
}