# lib/api-zod/src/generated/types/tokenInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArcMeme.fun API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TokenInput {
  /**
   * @minLength 1
   * @maxLength 50
   */
  name: string;
  /**
   * @minLength 1
   * @maxLength 10
   */
  ticker: string;
  description: string;
  website?: string;
  twitter?: string;
  telegram?: string;
  logoColor?: string;
  logoImage?: string;
  totalSupply?: number;
  contractAddress?: string;
  creatorAddress?: string;
}