# lib/api-zod/src/generated/types/tokenMarketInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArcMeme.fun API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TokenMarketInputMarketType } from "./tokenMarketInputMarketType";

export interface TokenMarketInput {
  marketType: TokenMarketInputMarketType;
  /** @nullable */
  pairAddress: string | null;
  /** @nullable */
  routerAddress: string | null;
}