# lib/api-zod/src/generated/types/trade.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ArcMeme.fun API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TradeSide } from "./tradeSide";

export interface Trade {
  id: string;
  tokenId: string;
  pairAddress: string;
  txHash: string;
  logIndex: number;
  blockNumber: number;
  side: TradeSide;
  tokenAmount: number;
  wusdcAmount: number;
  executionPrice: number;
  traderAddress: string;
  timestamp: string;
}